iPad app related changes
Universal app

+++ Savoir/Savoir/HelperClasses/AWPagedArrayLoadingOperations/MenuItemLoadingOperation.h
@@ -10,6 +10,6 @@
 
 @interface MenuItemLoadingOperation : DataLoadingOperation
 
-- (instancetype)initWithIndexes:(NSIndexSet *)indexes storeId:(long)storeId menuPOSId:(long)menuPOSId;
+- (instancetype)initWithIndexes:(NSIndexSet *)indexes storeId:(long long)storeId menuPOSId:(int)menuPOSId;
 
 @end
