+++ Savoir/Savoir/HelperClasses/AWPagedArrayLoadingOperations/ReviewLoadingOperation.m
@@ -22,7 +22,7 @@
 @implementation ReviewLoadingOperation
 @synthesize bDataLoaded = _bDataLoaded;
 
-- (instancetype)initWithIndexes:(NSIndexSet *)indexes storeId:(long)storeId
+- (instancetype)initWithIndexes:(NSIndexSet *)indexes storeId:(long long)storeId
 {
     
     self = [super initWithIndexes:indexes];
