+++ Savoir/Savoir/ViewControllers/Chat/Utilities/pushnotification.m
@@ -48,7 +48,7 @@ void ParsePushUserResign(void)
 }
 
 //-------------------------------------------------------------------------------------------------------------------------------------------------
-void SendPushNotification(long roomId, long storeId, NSString *text)
+void SendPushNotification(long long roomId, long long storeId, NSString *text)
 //-------------------------------------------------------------------------------------------------------------------------------------------------
 {
 //    __weak __typeof(self) weakSelf = self;
@@ -75,7 +75,7 @@ void SendPushNotification(long roomId, long storeId, NSString *text)
              PFQuery *queryInstallation = [PFInstallation query];
              [queryInstallation whereKey:PF_INSTALLATION_USER matchesQuery:query];
              
-             NSDictionary *pushContent = @{@"alert": text, @"TYPE":@"CHAT", @"CHAT_ROOMID":[NSString stringWithFormat:@"%ld", roomId], @"CHAT_STOREID":[NSString stringWithFormat:@"%ld", storeId]};
+             NSDictionary *pushContent = @{@"alert": text, @"TYPE":@"CHAT", @"CHAT_ROOMID":[NSString stringWithFormat:@"%lld", roomId], @"CHAT_STOREID":[NSString stringWithFormat:@"%lld", storeId]};
              
              PFPush *push = [[PFPush alloc] init];
              [push setQuery:queryInstallation];
@@ -85,13 +85,13 @@ void SendPushNotification(long roomId, long storeId, NSString *text)
               {
                   if (error != nil)
                   {
-                      NSLog(@"SendPushNotification send error:%ld, %@.", error.code, error.debugDescription);
+                      NSLog(@"SendPushNotification send error:%ld, %@.", (long)error.code, error.debugDescription);
                   }
               }];
          }
          else
          {
-             NSLog(@"queryForGetChatUsersForRoomWithRoomId error:%ld, %@.", error.code, error.debugDescription);
+             NSLog(@"queryForGetChatUsersForRoomWithRoomId error:%ld, %@.", (long)error.code, error.debugDescription);
          }
      }];
 }
